// TPT/old_version/Customer_Web/app/lib/api/apiTypes.ts
export type FeedbackStatus = "FR" | "FN" | "N"; // FR: 답변 읽음, FN: 답변 안 읽음, N: 답변 없음

/** 투자 타입 */
export type InvestmentType = "SWING" | "DAY" | "SCALPING";

/** 코스 상태 */
export type CourseStatus = "BEFORE_COMPLETION" | "PENDING_COMPLETION" | "AFTER_COMPLETION";
// ...
export type Position = "LONG" | "SHORT";

/** 진입 타점 */
export type EntryPoint = "REVERSE" | "PULL_BACK" | "BREAK_OUT";

/** 등급 */
export type Grade = "S_PLUS" | "S" | "A" | "B" | "NONE";
// ...
export interface DayFeedbackRequestDetailResponseDTO {
	id: number;
	createdAt: string;
	investmentType: InvestmentType;
	courseStatus: CourseStatus;
	membershipLevel: "BASIC" | "PREMIUM";
	feedbackYear: number;
	feedbackMonth: number;
	feedbackWeek: number;
	feedbackRequestDate: string;
	status: FeedbackStatus;
	isBestFeedback: boolean;
	updatedAt: string;
	category: string;
	positionHoldingTime: string;
	screenshotImageUrls: string[];
	riskTaking: number;
	leverage: number;
	positionStartDate: string;
	positionEndDate: string;
	position: Position;
	positionStartReason: string;
	positionEndReason: string;
	trainerFeedbackRequestContent: string;
	directionFrameExists: boolean;
	directionFrame?: string;
	mainFrame: string;
	subFrame: string;
	trendAnalysis: string;
	pnl: number;
	rnr: number;
	entryPoint1: EntryPoint;
	grade: Grade;
	entryPoint2?: string;
	tradingReview: string;
}

/** 스켈핑 트레이딩 피드백 요청 상세 응답 */
export interface ScalpingFeedbackRequestDetailResponseDTO {
	id: number;
	createdAt: string;
	investmentType: InvestmentType;
	courseStatus: CourseStatus;
	membershipLevel: "BASIC" | "PREMIUM";
	feedbackYear: number;
	feedbackMonth: number;
	feedbackWeek: number;
	feedbackRequestDate: string;
	status: FeedbackStatus;
	isBestFeedback: boolean;
	updatedAt: string;
	category: string;
	positionHoldingTime: string;
	screenshotImageUrls: string[];
	riskTaking: number;
	leverage: number;
	position: Position;
	pnl: number;
	rnr: number;
	operatingFundsRatio: number;
	entryPrice: number;
	exitPrice: number;
	settingStopLoss: number;
	settingTakeProfit: number;
	positionStartReason: string;
	positionEndReason: string;
	tradingReview: string;
}

/** 스윙 트레이딩 피드백 요청 상세 응답 */
export interface SwingFeedbackRequestDetailResponseDTO {
	id: number;
	createdAt: string;
	investmentType: InvestmentType;
	courseStatus: CourseStatus;
	membershipLevel: "BASIC" | "PREMIUM";
	feedbackRequestDate: string;
	status: FeedbackStatus;
	feedbackYear: number;
	feedbackMonth: number;
	feedbackWeek: number;
	isBestFeedback: boolean;
	updatedAt: string;
	category: string;
	positionHoldingTime: string;
	screenshotImageUrls: string[];
	riskTaking: number;
	leverage: number;
	positionStartDate: string;
	positionEndDate: string;
	position: Position;
	positionStartReason: string;
	positionEndReason: string;
	trainerFeedbackRequestContent: string;
	directionFrame: string;
	mainFrame: string;
	subFrame: string;
	trendAnalysis: string;
	pnl: number;
	rnr: number;
	entryPoint1: EntryPoint;
	grade: Grade;
	entryPoint2?: string;
	entryPoint3?: string;
	tradingReview: string;
}